pyproject.toml
[project]
name = "process-bot"
version = "0.1.0"
description = "Discord recruiting process tracker bot with analytics API and dashboard."
requires-python = ">=3.10"
dependencies = [
  "discord.py>=2.4.0,<3.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "jinja2>=3.1.0,<4.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "pydantic-settings>=2.5.0,<3.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["process_bot*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
